add RGBColor class and overall working implementation
This patchset is way too big but I got an implementation roll. This adds
an RGBColor class and predefined colors and the usage of the colors by
the LEDBinding and BlinkStick classes. This also has code for all the
communication and reports for the BlinkStick, minus support for
channels, but I'll need to see/get an example of how channels work. The
device I have doesn't support them.